+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './Components/App/App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './Services/WebVitals/reportWebVitals';
 
 ReactDOM.render(
   <React.StrictMode>
